Add guard clause

+++ tools/fetch_solution_candidates.py
@@ -20,6 +20,8 @@ def parse_args():
     """
     parser = argparse.ArgumentParser(description='Fetch competition metadata.')
     parser.add_argument('-s', '--slug', required=True, help='Competition slug')
+    parser.add_argument('-o', '--overwrite', action='store_false',
+                        help='If specified, overwrite an existing file.')
     return parser.parse_args()
 
 
@@ -67,6 +69,8 @@ def main():
     soup = make_soup(html)
     candidates = get_solution_candidates(soup)
     save_path = os.path.join('solution-candidates', f'{args.slug}.json')
+    if not args.overwrite and os.path.exists(save_path):
+        raise IOError('Could not save to', save_path, 'because the file already exists.')
     to_json(candidates, save_path)
     print('Saved to:', save_path)
